+++ pybind11/ccPolylinePy.cpp
@@ -27,6 +27,7 @@
 #include <ccShiftedObject.h>
 #include <GenericIndexedCloudPersist.h>
 #include <ccPointCloud.h>
+#include <CCGeom.h>
 #include "ccPolylinePy_DocStrings.hpp"
 
 #include <vector>
@@ -148,13 +149,63 @@ ccPointCloud* samplePoints_py( ccPolyline& self,
     return self.samplePoints(densityBased, samplingParameter, withRGB);
 }
 
+ccPolyline* createPolylineFromCoordinatesArray3(std::vector<CCVector3> points, bool isClosed)
+{
+    ccPointCloud* vertices = new ccPointCloud("vertices");
+    if (!vertices->reserve(points.size()))
+    {
+        delete vertices;
+        vertices = nullptr;
+        return nullptr;
+    }
+    for (auto point : points)
+    {
+        vertices->addPoint(point);
+    }
+    ccPolyline *polyline = new ccPolyline(vertices);
+    if (!polyline->reserve(vertices->size()))
+    {
+        delete polyline;
+        polyline = nullptr;
+        return polyline;
+    }
+    polyline->addPointIndex(0, vertices->size());
+    polyline->setClosed(isClosed);
+    return polyline;
+}
+
+ccPolyline* createPolylineFromCoordinatesArray2(std::vector<CCVector2> points, bool isClosed)
+{
+    ccPointCloud* vertices = new ccPointCloud("vertices");
+    if (!vertices->reserve(points.size()))
+    {
+        delete vertices;
+        vertices = nullptr;
+        return nullptr;
+    }
+    for (auto point : points)
+    {
+        vertices->addPoint(CCVector3(point[0], point[1], 0.));
+    }
+    ccPolyline *polyline = new ccPolyline(vertices);
+    if (!polyline->reserve(vertices->size()))
+    {
+        delete polyline;
+        polyline = nullptr;
+        return polyline;
+    }
+    polyline->addPointIndex(0, vertices->size());
+    polyline->setClosed(isClosed);
+    return polyline;
+}
+
 void export_ccPolyline(py::module &m0)
 {
 
 	py::class_<CCCoreLib::Polyline, CCCoreLib::ReferenceCloud,
                std::unique_ptr<CCCoreLib::Polyline, py::nodelete>> (m0, "Polyline",
 			   ccPolylinePy_Polyline_doc)
-		 .def(py::init<CCCoreLib::GenericIndexedCloudPersist*>())
+		 .def(py::init<CCCoreLib::GenericIndexedCloudPersist*>(), ccPolylinePy_PolylineCtor1_doc)
          .def("isClosed", &ccPolyline::isClosed, ccPolylinePy_isClosed_doc)
 	     .def("setClosed", &ccPolyline::setClosed, ccPolylinePy_setClosed_doc)
 		 ;
@@ -162,7 +213,9 @@ void export_ccPolyline(py::module &m0)
     py::class_<ccPolyline,CCCoreLib::Polyline, ccShiftedObject,
                std::unique_ptr<ccPolyline, py::nodelete> >(m0, "ccPolyline",
     	       ccPolylinePy_ccPolyline_doc)
-        .def(py::init<CCCoreLib::GenericIndexedCloudPersist*>())
+        .def(py::init<CCCoreLib::GenericIndexedCloudPersist*>(), ccPolylinePy_ccPolylineCtor1_doc)
+        .def(py::init(&createPolylineFromCoordinatesArray2), ccPolylinePy_ccPolylineCtor2_doc)
+        .def(py::init(&createPolylineFromCoordinatesArray3), ccPolylinePy_ccPolylineCtor3_doc)
         .def("computeLength", &ccPolyline::computeLength, ccPolylinePy_computeLength_doc)
         .def("getName", &ccPolyline::getName, ccPolylinePy_getName_doc)
         .def("is2DMode", &ccPolyline::is2DMode, ccPolylinePy_is2DMode_doc)