+++ pybind11/ccPrimitivesPy_DocStrings.hpp
@@ -30,16 +30,28 @@ Transformation (M) is composed by a rotation (R) and a translation (T):
 
 Available operators: ``+= -= *= *``
 
-Default constructor (no parameter) gives identity.
+See __init__ methods below for constructors.
+)";
 
-Alternate constructor 1:
+const char* ccPrimitivesPy_ccGLMatrix_ctor1_doc= R"(
+Default constructor (no parameter) gives identity.
+)";
+
+const char* ccPrimitivesPy_ccGLMatrix_ctor2_doc= R"(
+)";
+
+const char* ccPrimitivesPy_ccGLMatrix_ctor3_doc= R"(
+Alternate constructor
 
 :param tuple X: 3 first elements of the 1st column (last one is 0)
 :param tuple Y: 3 first elements of the 2nd column (last one is 0)
 :param tuple Z: 3 first elements of the 3rd column (last one is 0)
 :param tuple Tr: 3 first elements of the last column (last one is 1)
+)";
 
-Alternate constructor 2:
+const char* ccPrimitivesPy_ccGLMatrix_ctor4_doc= R"(
+Alternate constructor
+
 :param list data: 16 elements of the 4x4 matrix, stored in column major order (see :py:meth:`data`)
 )";
 
@@ -51,17 +63,7 @@ Transformation (M) is composed by a rotation (R) and a translation (T):
 
 Available operators: ``+= -= *= *``
 
-Default constructor (no parameter) gives identity.
-
-Alternate constructor 1:
-
-:param tuple X: 3 first elements of the 1st column (last one is 0)
-:param tuple Y: 3 first elements of the 2nd column (last one is 0)
-:param tuple Z: 3 first elements of the 3rd column (last one is 0)
-:param tuple Tr: 3 first elements of the last column (last one is 1)
-
-Alternate constructor 2:
-:param list data: 16 elements of the 4x4 matrix, stored in column major order (see :py:meth:`data`)
+See __init__ methods below for constructors.
 )";
 
 const char* ccPrimitivesPy_ccGLMatrixParams1_doc= R"(
@@ -115,6 +117,12 @@ Returns Height.
 :rtype: double
 )";
 
+const char* ccPrimitivesPy_setHeight_doc= R"(
+Set or update height value
+
+:param double height: new height value
+)";
+
 const char* ccPrimitivesPy_getBottomRadius_doc= R"(
 Returns bottom radius.
 
@@ -122,11 +130,29 @@ Returns bottom radius.
 :rtype: double
 )";
 
+const char* ccPrimitivesPy_setBottomRadius_doc= R"(
+Set or update bottom radius.
+
+:param double radius: new bottom radius value
+)";
+
 const char* ccPrimitivesPy_getTopRadius_doc= R"(
 Returns top radius.
 
 :return: top radius
 :rtype: double
+)";
+
+const char* ccPrimitivesPy_setTopRadius_doc= R"(
+Set or update top radius.
+
+:param double radius: new top radius value
+)";
+
+const char* ccPrimitivesPy_setRadius_doc= R"(
+Set or update radius.
+
+:param double radius: new radius value
 )";
 
 const char* ccPrimitivesPy_getBottomCenter_doc= R"(
@@ -288,6 +314,14 @@ const char* ccPrimitivesPy_ccBox_doc= R"(
 Box dimensions axis along each dimension are defined in a single 3D vector.
 
 A box is in fact composed of 6 planes (ccPlane).
+)";
+
+const char* ccPrimitivesPy_ccBox_ctor1_doc= R"(
+Simplified ccBox constructor, creates a box with null dimensions. (Not for a standard usage).
+)";
+
+const char* ccPrimitivesPy_ccBox_ctor2_doc= R"(
+ccBox Constructor
 
 :param tuple dims: box dimensions
 :param ccGLMatrix,optional transMat: 3D transformation, default 0
@@ -302,8 +336,22 @@ Get the box dimensions.
 :rtype: tuple dimensions
 )";
 
+const char* ccPrimitivesPy_ccBox_setDimensions_doc= R"(
+Set or update the box dimensions.
+
+:param tuple dims: box dimensions
+)";
+
 const char* ccPrimitivesPy_ccCone_doc= R"(
 Cone axis corresponds to the 'Z' dimension by default.
+)";
+
+const char* ccPrimitivesPy_ccCone_ctor1_doc= R"(
+Simplified ccCone constructor, creates a cone with null dimensions. (Not for a standard usage).
+)";
+
+const char* ccPrimitivesPy_ccCone_ctor2_doc= R"(
+ccCone constructor.
 
 :param float bottomRadius: cone bottom radius
 :param float topRadius: cone top radius
@@ -314,12 +362,21 @@ Cone axis corresponds to the 'Z' dimension by default.
        (can be set afterwards with ccDrawableObject::setGLTransformation)
 :param str,optional name: name, default 'Cone'
 :param int,optional precision: drawing precision (angular step = 360/precision), default 24
-:param int,optional uniqueID:  unique ID (handle with care))";
+:param int,optional uniqueID:  unique ID (handle with care)
+)";
 
 const char* ccPrimitivesPy_ccCylinder_doc= R"(
 Cylinder axis corresponds to the 'Z' dimension.
 
 Internally represented by a cone with the same top and bottom radius.
+)";
+
+const char* ccPrimitivesPy_ccCylinder_ctor1_doc= R"(
+Simplified ccCylinder constructor, with null dimensions. (Not for a standard usage).
+)";
+
+const char* ccPrimitivesPy_ccCylinder_ctor2_doc= R"(
+ccCylinder constructor
 
 :param float radius: cylinder radius
 :param float height: cylinder height (transformation should point to the axis center)
@@ -327,16 +384,29 @@ Internally represented by a cone with the same top and bottom radius.
        (can be set afterwards with ccDrawableObject::setGLTransformation)
 :param str,optional name: name, default 'Cylinder'
 :param int,optional precision: drawing precision (angular step = 360/precision), default 24
-:param int,optional uniqueID: unique ID (handle with care))";
+:param int,optional uniqueID: unique ID (handle with care)
+)";
 
 const char* ccPrimitivesPy_ccPlane_doc= R"(
-Plane normal corresponds to 'Z' dimension.
+Default plane normal corresponds to 'Z' dimension.
+a ccPlane is defined with widths along 'X' and 'Y' dimensions. 
+
+See __init__ methods below for constructors.
+)";
+
+const char* ccPrimitivesPy_ccPlane_ctor1_doc= R"(
+Simplified ccPlane constructor, with null dimensions. (Not for a standard usage).
+)";
+
+const char* ccPrimitivesPy_ccPlane_ctor2_doc= R"(
+ccPlane constructor
 
 :param float xWidth: plane width along 'X' dimension
 :param float yWidth: plane width along 'Y' dimension
 :param ccGLMatrix,optional transMat: (default 0) 3D transformation
        (can be set afterwards with ccDrawableObject::setGLTransformation)
-:param str,optional name: name, default 'Plane')";
+:param str,optional name: name, default 'Plane'
+)";
 
 const char* ccPrimitivesPy_ccPlane_Fit_doc= R"(
 Fits a plane primitive on a cloud.
@@ -367,8 +437,24 @@ Returns the normal of the plane.
 :return: plane normal
 :rtype: CCVector3)";
 
+const char* ccPrimitivesPy_ccPlane_setXWidth_doc= R"(
+Set or update plane width along 'X' dimension.
+
+:param float xWidth: plane width along 'X' dimension
+)";
+
+const char* ccPrimitivesPy_ccPlane_setYWidth_doc= R"(
+Set or update plane width along 'Y' dimension.
+
+:param float yWidth: plane width along 'Y' dimension
+)";
+
 const char* ccPrimitivesPy_ccQuadric_doc= R"(
 Quadric orthogonal dimension is 'Z' by default.
+)";
+
+const char* ccPrimitivesPy_ccQuadric_ctor_doc= R"(
+ccQuadric constructor
 
 :param tuple minCorner: min corner of the 'representation' base area
 :param tuple maxCorner: max corner of the 'representation' base area
@@ -377,14 +463,26 @@ Quadric orthogonal dimension is 'Z' by default.
 :param ccGLMatrix,optional transMat: (default 0) 3D transformation
        (can be set afterwards with ccDrawableObject::setGLTransformation)
 :param str,optional name: optional name, default 'Quadric'
-:param int,optional precision: optional drawing precision, default 24)";
+:param int,optional precision: optional drawing precision, default 24
+)";
 
 const char* ccPrimitivesPy_ccSphere_doc= R"(
+Sphere primitive, mainly defined by its radius and drawing precision.
+)";
+
+const char* ccPrimitivesPy_ccSphere_ctor1_doc= R"(
+Simplified ccSphere constructor with null radius.  (Not for a standard usage).
+)";
+
+const char* ccPrimitivesPy_ccSphere_ctor2_doc= R"(
+ccSphere constructor
+
 :param float radius: sphere radius.
 :param ccGLMatrix,optional transMat: (default 0) 3D transformation (can be set afterwards with ccDrawableObject::setGLTransformation)
 :param str,optional name: optional name, default 'Sphere'
 :param int,optional precision: optional drawing precision (angular step = 360/precision), default 24
-:param int,optional uniqueID: optional unique ID (handle with care))";
+:param int,optional uniqueID: optional unique ID (handle with care)
+)";
 
 const char* ccPrimitivesPy_ccSphere_getRadius_doc= R"(
 Get the sphere radius.
@@ -393,8 +491,23 @@ Get the sphere radius.
 :rtype: float
 )";
 
+const char* ccPrimitivesPy_ccSphere_setRadius_doc= R"(
+Set the sphere radius.
+
+:param float radius: sphere radius.
+)";
+
 const char* ccPrimitivesPy_ccTorus_doc= R"(
 Torus is defined in the XY plane by default.
+Its section can be rectangular or round.
+)";
+
+const char* ccPrimitivesPy_ccTorus_ctor1_doc= R"(
+Simplified ccTorus constructor with null dimensions. (Not for a standard usage).
+)";
+
+const char* ccPrimitivesPy_ccTorus_ctor2_doc= R"(
+ccTorus constructor
 
 :param float insideRadius: inside radius
 :param float outsideRadius: outside radius
@@ -407,12 +520,21 @@ Torus is defined in the XY plane by default.
 :param int,optional precision: drawing precision
        (main loop angular step = 360/(4*precision),
        circular section angular step = 360/precision), default 24
-:param int,optional uniqueID: unique ID (handle with care))";
+:param int,optional uniqueID: unique ID (handle with care)
+)";
 
 const char* ccPrimitivesPy_ccDish_doc= R"(
 Either a section of a sphere, or half of an ellipsoid!
 
 The origin of the Dish is at the centre of the base, and the Z-axis lies along the normal to the base.
+)";
+
+const char* ccPrimitivesPy_ccDish_ctor1_doc= R"(
+Simplified ccDish constructor with null dimensions. (Not for a standard usage).
+)";
+
+const char* ccPrimitivesPy_ccDish_ctor2_doc= R"(
+ccDish constructor
 
 :param float radius: base radius
 :param float height: maximum height of dished surface above base
@@ -421,7 +543,8 @@ The origin of the Dish is at the centre of the base, and the Z-axis lies along t
 :param ccGLMatrix,optional transMat: (default 0) 3D transformation
        (can be set afterwards with ccDrawableObject::setGLTransformation)
 :param str,optional name: name, default 'Dish'
-:param int,optional precision: drawing precision (angular step = 360/precision), default 24)";
+:param int,optional precision: drawing precision (angular step = 360/precision), default 24
+)";
 
 const char* ccPrimitivesPy_toString_doc= R"(
 String version of the transformer coefficients